+++ service/controller/v13/encrypter/kms/spec.go
@@ -1,5 +1,11 @@
 package kms
 
+import "github.com/aws/aws-sdk-go/service/kms"
+
 const (
 	pendingDeletionWindow = 7
 )
+
+type KMSClient interface {
+	Encrypt(*kms.EncryptInput) (*kms.EncryptOutput, error)
+}
